# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "KepRoop API Documentation",
    "version": "1.0"
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    },
    {
      "CookieAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Prefix with 'Api-Key ' or 'Bearer kp_'"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "Browser Session Cookie (Automatically used if logged in)"
      }
    }
  },
  "paths": {
...
    "/api/auth/api-keys": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "List API keys",
        "description": "List all API keys for the authenticated user.",
        "responses": {
          "200": {
            "description": "List of API keys"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Create API key",
        "description": "Create a new API key",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created API key (only shown once)"
          }
        }
      }
    },
...
    "/api/images/{id}/download-url": {
      "get": {
        "tags": [
          "Images"
        ],
        "summary": "Get download URL",
        "description": "Get a signed URL for downloading an image.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Download URL"
          }
        }
      }
    },
    "/api/images/{id}": {
      "get": {
        "tags": [
          "Images"
        ],
        "summary": "Get image details",
        "description": "Get image download URLs and metadata.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image details"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Images"
        ],
        "summary": "Delete image",
        "description": "Soft delete an image.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image deleted"
          }
        }
      },
      "patch": {
        "tags": [
          "Images"
        ],
        "summary": "Restore image",
        "description": "Restore a soft-deleted image.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "action"
                ],
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "restore"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image restored"
          }
        }
      }
    },
    "/api/images/bulk": {
      "post": {
        "tags": [
          "Images"
        ],
        "summary": "Bulk operations",
        "description": "Perform bulk delete, download, or move operations.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "action",
                  "imageIds",
                  "albumId"
                ],
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "delete",
                      "download",
                      "move"
                    ]
                  },
                  "imageIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "albumId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operation successful"
          }
        }
      }
    },
    "/api/images/register": {
      "post": {
        "tags": [
          "Images"
        ],
        "summary": "Register uploaded image",
        "description": "Register an image after successful client-side upload.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "albumId",
                  "s3KeyOriginal"
                ],
                "properties": {
                  "albumId": {
                    "type": "string"
                  },
                  "s3KeyOriginal": {
                    "type": "string"
                  },
                  "filename": {
                    "type": "string"
                  },
                  "size": {
                    "type": "integer"
                  },
                  "width": {
                    "type": "integer"
                  },
                  "height": {
                    "type": "integer"
                  },
                  "folderId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image registered"
          }
        }
      }
    },
...
  },
...
}